Cogs Added
All functions which talk to mongodb moved to scheduleFetcher. Function calls to linkSender functions moved to linkSender.py
