+++ cogs/__init__.py
@@ -1,0 +1,2 @@
+def setup(bot):
+    print("Cogs Loading....")
